feat(cache manager): add webpack config to cache manager
add webpack config to cache manager
