+++ apps/cache-manager/tsconfig.json
@@ -14,13 +14,17 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": true,
+    "noEmit": false,
     "jsx": "react-jsx"
   },
   "ts-node": {
     // "esm": true,
     "compilerOptions": {
-      "module": "CommonJS"
+      "module": "CommonJS",
+      "sourceMap": true
     }
-}
+  },
+  "include": [
+    "./**/*.ts"
+  ]
 }
