+++ apps/cache-manager/webpack.config.cjs
@@ -1,0 +1,24 @@
+const path = require('path')
+
+const {NODE_ENV = 'production'} = process.env
+
+module.exports = {
+  entry: './src/server.ts',
+  mode: NODE_ENV,
+  target: 'node',
+  output: {
+    path: path.resolve(__dirname, 'build'),
+    filename: 'index.js',
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        use: ['ts-loader'],
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['.ts', '.js'],
+  },
+}
