+++ package.json
@@ -185,11 +185,13 @@
     "source-map-loader": "^3.0.0",
     "ssl-root-cas": "^1.3.1",
     "style-loader": "^3.3.1",
+    "supertest": "^6.2.4",
     "tailwindcss": "^3.0.2",
     "terser-webpack-plugin": "^5.2.5",
+    "ts-loader": "^9.4.1",
     "ts-node": "^10.4.0",
     "typedoc": "^0.22.10",
-    "typescript": "^4.5.4",
+    "typescript": "^4.8.4",
     "url-loader": "^4.1.1",
     "web-vitals": "^2.1.2",
     "webpack": "^5.64.4",
